Refactor common multisets from ObservationResultsStore (#4343)
This PR only has refactors

Pulls the BiomassDetails and the per-substratum/stratum/site species
multisets — plus the helpers they share (plotHasCompletedObservations,
speciesMultiset, latestLiveField) — out of ObservationResultsStore into
two new files:

- BiomassDetailsMultiset.kt
- ObservationMultisets.kt

ObservationResultsStore continues to call them by name. No behaviour
change; this is just a move-and-rewire so the same multisets can be
reused by a forthcoming V2 results store.
